chore(session): fix NPE when session to delete is already gone

+++ core/src/main/java/de/arbeitsagentur/opdt/keycloak/cassandra/userSession/persistence/CassandraUserSessionRepository.java
@@ -135,6 +135,10 @@ public class CassandraUserSessionRepository implements UserSessionRepository {
 
   @Override
   public void deleteUserSession(UserSession session) {
+    if (session == null) {
+      return;
+    }
+    
     dao.deleteUserSession(session);
 
     // Attributes
